requirements.txt
Flask==3.0.3
Flask-SocketIO==5.4.1
mysql-connector-python==8.4.0

